refactor: scripts to use ESM
use ".mts" extension (setting "type": "module" breaks Angular prerender)
specify extension when importing in scripts
move utils into "src" so it's included as part of tsconfig.json
update tsconfig.json to use nodenext for ESM enablement
update tsconfig.json to output to "dist" for debug js compilation
update tsconfig.json to tell ts-node to enable ESM
update utils to use ESM (isMain, __dirname subst)
update import of metadata.js to use CommonJs compatibility

(cherry picked from commit adcd253ff50b1ab81a891f5602052e673a3b0029)
